+++ src/CHANGELOG.rst
@@ -5,8 +5,8 @@ v4.0dev25 = New Dev Release (???)
 ---------------------------
 - Fixed /cache command.
 - Added mention method to command object.
-- Added 'is_child' and 'parent_cmd' property in command object.
-
+- Added 'is_child' and 'parent_cmd' property to command object.
+- Added dropdown view function in goldy bot utils.
 
 v4.0dev24 = New Dev Release (01/10/2022)
 ---------------------------